Move Azure DevOps helpers from common/ to go_modules/
Relocate configure_go_for_azure_devops and related methods from
SharedHelpers (common/) into a dedicated GoModules::AzureDevOpsHelper
module. These are Go-specific concerns and should not live in the
shared common library.
